refactor: extract raycast effect into separate file and fix wall turning
Split the raycaster effect from the monolithic shenanigans.go (7500+ lines)
into its own effect_raycast.go file to improve readability and maintainability.

Fixed the "nudge" when turning at walls by replacing the instant multi-rotation
burst with a committed-direction state machine: once a turn direction is chosen,
it persists until the path ahead is fully clear, eliminating the back-and-forth
hesitation.

Also fixed the plugin build in Taskfile.build.yml to use package directories
(./plugins/<name>/) instead of single files, which is required for multi-file
plugin packages and matches what .goreleaser.yaml already does.
